feat: improve editor, implement dropdown, enum editing and a lot of ffi throughout the project

+++ fruits_ecs/src/data/entity/unique_components_set.rs
@@ -28,7 +28,7 @@ impl PartialEq for UniqueComponentsSet {
             return false;
         }
 
-        for i in 0..(self.components.len() as usize) {
+        for i in 0..(self.components.len()) {
             if self.components[i] != other.components[i] {
                 return false;
             }
